+++ app/src/main/java/com/cmc15th/pluv/ui/home/migrate/direct/DisplayMigrationPathScreen.kt
@@ -32,8 +32,8 @@ import com.cmc15th.pluv.core.designsystem.component.TopBarWithProgress
 import com.cmc15th.pluv.core.designsystem.theme.Title1
 import com.cmc15th.pluv.domain.model.PlayListApp
 import com.cmc15th.pluv.ui.home.getAppNameRes
-import com.cmc15th.pluv.ui.home.migrate.component.FetchPlaylistLoadingIcon
-import com.cmc15th.pluv.ui.home.migrate.component.PreviousOrMigrateButton
+import com.cmc15th.pluv.ui.home.migrate.common.component.FetchPlaylistLoadingIcon
+import com.cmc15th.pluv.ui.home.migrate.common.component.PreviousOrMigrateButton
 
 @Composable
 fun DisplayMigrationPathScreen(
