+++ app/src/main/java/com/cmc15th/pluv/ui/home/migrate/direct/SelectDestinationAppScreen.kt
@@ -29,7 +29,7 @@ import com.cmc15th.pluv.core.designsystem.theme.Title1
 import com.cmc15th.pluv.domain.model.PlayListApp
 import com.cmc15th.pluv.ui.home.getAppNameRes
 import com.cmc15th.pluv.ui.home.getSelectedIconRes
-import com.cmc15th.pluv.ui.home.migrate.component.PreviousOrMigrateButton
+import com.cmc15th.pluv.ui.home.migrate.common.component.PreviousOrMigrateButton
 
 @Composable
 fun SelectDestinationAppScreen(
