+++ app/src/main/java/com/cmc15th/pluv/ui/home/migrate/direct/SelectSourceAppScreen.kt
@@ -31,7 +31,7 @@ import com.cmc15th.pluv.core.designsystem.theme.Title3
 import com.cmc15th.pluv.domain.model.PlayListApp
 import com.cmc15th.pluv.ui.home.getAppIconRes
 import com.cmc15th.pluv.ui.home.getAppNameRes
-import com.cmc15th.pluv.ui.home.migrate.component.PreviousOrMigrateButton
+import com.cmc15th.pluv.ui.home.migrate.common.component.PreviousOrMigrateButton
 
 @Composable
 fun SelectSourceAppScreen(
